Refactor splitPath into router utils
