+++ lib/client-api/src/index.js
@@ -1,5 +1,5 @@
 import ClientApi, { defaultDecorateStory } from './client_api';
-import StoryStore, { splitPath } from './story_store';
+import StoryStore from './story_store';
 import ConfigApi from './config_api';
 import subscriptionsStore from './subscriptions_store';
 import pathToId from './pathToId';
@@ -13,7 +13,6 @@ export {
   subscriptionsStore,
   defaultDecorateStory,
   pathToId,
-  splitPath,
   getQueryParams,
   getQueryParam,
 };
