+++ lib/client-api/src/story_store.js
@@ -21,17 +21,6 @@ const toKey = input =>
 
 const toChild = it => ({ ...it });
 
-export const splitPath = (path, { rootSeparator, groupSeparator }) => {
-  const [root, remainder] = path.split(rootSeparator, 2);
-  const groups = (remainder || path).split(groupSeparator).filter(i => !!i);
-
-  // when there's no remainder, it means the root wasn't found/split
-  return {
-    root: remainder ? root : null,
-    groups,
-  };
-};
-
 let count = 0;
 
 function getId() {
